Filter the genomes of a tidygenomes object

This function applies dplyr::filter() to the genomes table of a tidygenomes object and subsequently filters all other components of the tidygenomes object to retain only the requested genomes. It is also possible to filter on variables of the phylogroups table, if present.

filter_genomes(tg, ...)

Arguments

tg

A tidygenomes object

...

Arguments passed on to dplyr::filter

.data

A data frame, data frame extension (e.g. a tibble), or a lazy data frame (e.g. from dbplyr or dtplyr). See Methods, below, for more details.

.by

<tidy-select> Optionally, a selection of columns to group by for just this operation, functioning as an alternative to group_by(). For details and examples, see ?dplyr_by.

.preserve

Relevant when the .data input is grouped. If .preserve = FALSE (the default), the grouping structure is recalculated based on the resulting data, otherwise the grouping is kept as is.

Value

A tidygenomes object

Examples

genomes <- data.frame(genome = c("genome 1", "genome 2"))
tg <- as_tidygenomes(genomes)
#> Creating tidygenomes object from genome table
tg <- tg %>% filter_genomes(genome == "genome 1")