Root tree given three tips

This function roots a phylogenetic tree given three tip labels.

root_tree.phylo(tree, tips, ...)

Arguments

tree

An object of class phylo

tips

Three tip labels

...

Arguments passed on to ape::root.phylo

phy

an object of class "phylo" or "multiPhylo".

outgroup

a vector of mode numeric or character specifying the new outgroup.

node

alternatively, a node number where to root the tree.

resolve.root

a logical specifying whether to resolve the new root as a bifurcating node.

interactive

if TRUE the user is asked to select the node by clicking on the tree which must be plotted.

edgelabel

a logical value specifying whether to treat node labels as edge labels and thus eventually switching them so that they are associated with the correct edges when using drawSupportOnEdges (see Czech et al. 2016).

Value

An object of class phylo

Details

Tips a, b and c define exactly one internal node in the unrooted tree. The tree will be rooted on the branch leading from this node to tip a.

By default, node/tip labels are interpreted as vertex labels, not edge labels. When the labels actually represent edge labels (e.g., branch support values), set edge_label to TRUE. This will make sure that the edge labels are handled correctly, i.e. moved to the correct nodes. In that case, when the tree is rooted on a non-tip branch, its edge label will be copied to both child nodes of the new root node.